# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Matador API
  description: Matador on-demand car rental API. Customer endpoints use JWT bearer
    tokens; admin endpoints use session cookies.
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/admin/zones:
    get:
      tags:
      - Admin-Settings
      summary: List zones
      description: All zones with GeoJSON boundaries.
      operationId: list_2
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ZoneResponse'
    post:
      tags:
      - Admin-Settings
      summary: Create zone
      description: Create a zone from a GeoJSON polygon.
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateZoneRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ZoneResponse'
  /api/admin/vehicles:
    get:
      tags:
      - Admin-Vehicles
      summary: List vehicles
      description: Paginated, filterable fleet list.
      operationId: list_3
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - AVAILABLE
          - RESERVED
          - IN_PREPARATION
          - EN_ROUTE_TO_CUSTOMER
          - WITH_CUSTOMER
          - EN_ROUTE_TO_WAREHOUSE
          - AWAITING_INSPECTION
          - IN_CLEANING
          - IN_MAINTENANCE
          - OUT_OF_SERVICE
          - RETIRED
      - name: classId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: zoneId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageVehicleResponse'
    post:
      tags:
      - Admin-Vehicles
      summary: Create vehicle
      description: Add a vehicle to the fleet.
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicleRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleResponse'
  /api/admin/vehicles/{id}/unlock:
    post:
      tags:
      - Admin-Vehicles
      summary: Unlock vehicle
      description: Send an unlock command via telematics.
      operationId: unlock
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommandResponse'
  /api/admin/vehicles/{id}/status:
    post:
      tags:
      - Admin-Vehicles
      summary: Transition status
      description: Validated vehicle status transition.
      operationId: transition
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransitionStatusRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleResponse'
  /api/admin/vehicles/{id}/lock:
    post:
      tags:
      - Admin-Vehicles
      summary: Lock vehicle
      description: Send a lock command via telematics.
      operationId: lock
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommandResponse'
  /api/admin/vehicle-classes:
    get:
      tags:
      - Admin-Vehicles
      summary: List vehicle classes
      description: All vehicle classes.
      operationId: list_4
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleClassResponse'
    post:
      tags:
      - Admin-Vehicles
      summary: Create vehicle class
      description: Create a vehicle class.
      operationId: create_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicleClassRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleClassResponse'
...
  /api/admin/zones/{id}:
    patch:
      tags:
      - Admin-Settings
      summary: Update zone
      description: Update fees, dropoff policy, or active flag.
      operationId: update_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateZoneRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ZoneResponse'
  /api/admin/vehicles/{id}:
    get:
      tags:
      - Admin-Vehicles
      summary: Vehicle detail
      description: Full vehicle detail.
      operationId: get_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleResponse'
    patch:
      tags:
      - Admin-Vehicles
      summary: Update vehicle
      description: Update mutable vehicle fields.
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVehicleRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleResponse'
  /api/admin/vehicle-classes/{id}:
    patch:
      tags:
      - Admin-Vehicles
      summary: Update vehicle class
      description: Update a vehicle class.
      operationId: update_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVehicleClassRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleClassResponse'
...
  /api/zones/contains:
    get:
      tags:
      - Customer-Vehicles
      summary: Zone containment
      description: Which zone (if any) contains a point.
      operationId: contains
      parameters:
      - name: lat
        in: query
        required: true
        schema:
          type: number
          format: double
      - name: lng
        in: query
        required: true
        schema:
          type: number
          format: double
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ZoneContainsResponse'
...
  /api/customer/vehicles/available:
    get:
      tags:
      - Customer-Vehicles
      summary: Available classes
      description: Bookable classes with counts and starting price for a window/location.
      operationId: available
      parameters:
      - name: pickupAt
        in: query
        required: true
        schema:
          type: string
          format: date-time
      - name: dropoffAt
        in: query
        required: true
        schema:
          type: string
          format: date-time
      - name: pickupLat
        in: query
        required: true
        schema:
          type: number
          format: double
      - name: pickupLng
        in: query
        required: true
        schema:
          type: number
          format: double
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvailableClassResponse'
...
components:
  schemas:
...
    CreateZoneRequest:
      required:
      - boundary
      - center
      - name
      - slug
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        boundary:
          $ref: '#/components/schemas/PolygonGeoJson'
        center:
          $ref: '#/components/schemas/PointGeoJson'
        outOfZoneDropoffFeeCents:
          type: integer
          format: int64
        outOfZoneDropoffAllowed:
          type: boolean
    PointGeoJson:
      type: object
      properties:
        type:
          type: string
        coordinates:
          type: array
          items:
            type: number
            format: double
    PolygonGeoJson:
      type: object
      properties:
        type:
          type: string
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              items:
                type: number
                format: double
    ZoneResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        boundary:
          $ref: '#/components/schemas/PolygonGeoJson'
        center:
          $ref: '#/components/schemas/PointGeoJson'
        outOfZoneDropoffFeeCents:
          type: integer
          format: int64
        outOfZoneDropoffAllowed:
          type: boolean
        active:
          type: boolean
    CreateVehicleRequest:
      required:
      - acquiredOn
      - classId
      - color
      - homeZoneId
      - licensePlate
      - licenseState
      - make
      - model
      - vin
      type: object
      properties:
        vin:
          type: string
        licensePlate:
          type: string
        licenseState:
          maxLength: 2
          minLength: 2
          type: string
        make:
          type: string
        model:
          type: string
        year:
          minimum: 1900
          type: integer
          format: int32
        color:
          type: string
        classId:
          type: string
          format: uuid
        homeZoneId:
          type: string
          format: uuid
        acquiredOn:
          type: string
          format: date
        telematicsProvider:
          type: string
        telematicsVehicleId:
          type: string
    VehicleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vin:
          type: string
        licensePlate:
          type: string
        licenseState:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
          format: int32
        color:
          type: string
        classId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - AVAILABLE
          - RESERVED
          - IN_PREPARATION
          - EN_ROUTE_TO_CUSTOMER
          - WITH_CUSTOMER
          - EN_ROUTE_TO_WAREHOUSE
          - AWAITING_INSPECTION
          - IN_CLEANING
          - IN_MAINTENANCE
          - OUT_OF_SERVICE
          - RETIRED
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        currentAddress:
          type: string
        odometerMiles:
          type: integer
          format: int32
        fuelChargePercent:
          type: integer
          format: int32
        rangeMiles:
          type: integer
          format: int32
        homeZoneId:
          type: string
          format: uuid
        telematicsProvider:
          type: string
        acquiredOn:
          type: string
          format: date
        retiredOn:
          type: string
          format: date
    CommandResponse:
      type: object
      properties:
        succeeded:
          type: boolean
        detail:
          type: string
    TransitionStatusRequest:
      required:
      - status
      type: object
      properties:
        status:
          type: string
          enum:
          - AVAILABLE
          - RESERVED
          - IN_PREPARATION
          - EN_ROUTE_TO_CUSTOMER
          - WITH_CUSTOMER
          - EN_ROUTE_TO_WAREHOUSE
          - AWAITING_INSPECTION
          - IN_CLEANING
          - IN_MAINTENANCE
          - OUT_OF_SERVICE
          - RETIRED
    CreateVehicleClassRequest:
      required:
      - drivetrain
      - name
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        seats:
          minimum: 1
          type: integer
          format: int32
        luggageCapacity:
          minimum: 0
          type: integer
          format: int32
        drivetrain:
          type: string
          enum:
          - ICE
          - HYBRID
          - EV
        baseDailyRateCents:
          minimum: 0
          type: integer
          format: int64
        sortOrder:
          type: integer
          format: int32
    VehicleClassResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        seats:
          type: integer
          format: int32
        luggageCapacity:
          type: integer
          format: int32
        drivetrain:
          type: string
          enum:
          - ICE
          - HYBRID
          - EV
        baseDailyRateCents:
          type: integer
          format: int64
        sortOrder:
          type: integer
          format: int32
        active:
          type: boolean
...
    UpdateZoneRequest:
      type: object
      properties:
        outOfZoneDropoffFeeCents:
          type: integer
          format: int64
        outOfZoneDropoffAllowed:
          type: boolean
        active:
          type: boolean
    UpdateVehicleRequest:
      type: object
      properties:
        color:
          type: string
        licensePlate:
          type: string
        notes:
          type: string
    UpdateVehicleClassRequest:
      type: object
      properties:
        description:
          type: string
        baseDailyRateCents:
          type: integer
          format: int64
        sortOrder:
          type: integer
          format: int32
        active:
          type: boolean
...
    ZoneContainsResponse:
      type: object
      properties:
        inZone:
          type: boolean
        zoneId:
          type: string
          format: uuid
        zoneName:
          type: string
...
    AvailableClassResponse:
      type: object
      properties:
        classId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        drivetrain:
          type: string
          enum:
          - ICE
          - HYBRID
          - EV
        seats:
          type: integer
          format: int32
        luggageCapacity:
          type: integer
          format: int32
        countAvailable:
          type: integer
          format: int64
        startingDailyRateCents:
          type: integer
          format: int64
        representativePhotoUrl:
          type: string
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
...
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean
...
    PageVehicleResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/VehicleResponse'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        empty:
          type: boolean
...
  securitySchemes:
    bearer-jwt:
      type: http
      description: Customer JWT access token.
      scheme: bearer
      bearerFormat: JWT
    session-cookie:
      type: apiKey
      description: Admin session cookie.
      name: SESSION
      in: cookie